+++ lib/generators/rodauth/templates/app/controllers/rodauth_controller.rb
@@ -1,3 +1,4 @@
 class RodauthController < ApplicationController
-  # used by Rodauth for rendering views and CSRF protection
+  # used by Rodauth for rendering views, CSRF protection, and running any
+  # registered action callbacks and rescue_from handlers
 end